+++ src/util/Path.cpp
@@ -43,7 +43,7 @@ namespace utl {
 	}
 
 	
-	std::wstring Path::compact(unsigned int max_char) const
+	std::wstring Path::compact(size_t max_char) const
 	{
 		std::vector<wchar_t> buffer(max_char + 1);
 		if (max_char > 0 && ::PathCompactPathEx(buffer.data(), to_string().c_str(), max_char, 0))
